Imlemented NetBuffer serialization system
Added basic framedata structs and hitbox systems.
Added more functionality to FixedPoint type (root function, floor/ceil function).
Implemented basic Input functionality (WIP).
Added manager classes for systems that didn't have one yet (not filled in yet).

+++ KraFight/KraFight/source/KraFight/Math/Vector2.cpp
@@ -78,6 +78,6 @@ Vector2 kra::Normalize(const Vector2 & Vec)
 kfloat kra::Length(const Vector2 & Vec)
 {
 	kfloat Ret = Vec.X * Vec.X + Vec.Y * Vec.Y;
-	Ret = sqrtf(Ret);
+	Ret = sqrt(Ret);
 	return Ret;
 }
